Add "https://" if instance URL misses it.
This patch adds a small function which tests for presence of either
"http://" or "https://" in the instance URL inserted by the user.

If none of them is found, it adds "https://" by default.

I noticed that somehow go-mastodon does not handle http->https redirect,
so consider this more of an hack just to not break UX.